Add IRC integration through KiwiIRC
I created a view, IrcView, and a template which wraps around the KiwiIRC
iframe. We need to add a link to that page in the navigation or on the
home page. We could also add a header on the IRC page with the compsoc
logo and a link to go back to the website.